+++ examples/mcxa/src/bin/adc_interrupt.rs
@@ -2,34 +2,31 @@
 #![no_main]
 
 use embassy_executor::Spawner;
-use embassy_mcxa_examples::init_adc_pins;
-use hal::adc::{LpadcConfig, TriggerPriorityPolicy};
+use embassy_time::{Duration, Ticker};
+use hal::adc::{Adc, InterruptHandler, LpadcConfig, TriggerPriorityPolicy};
+use hal::bind_interrupts;
 use hal::clocks::PoweredClock;
+use hal::clocks::config::Div8;
 use hal::clocks::periph_helpers::{AdcClockSel, Div4};
+use hal::config::Config;
 use hal::pac::adc1::cfg::{Pwrsel, Refsel};
-use hal::pac::adc1::cmdl1::{Adch, Mode};
+use hal::pac::adc1::cmdl1::Mode;
 use hal::pac::adc1::ctrl::CalAvgs;
-use hal::pac::adc1::tctrl::Tcmd;
-use hal::{InterruptExt, bind_interrupts};
+use hal::peripherals::ADC1;
 use {defmt_rtt as _, embassy_mcxa as hal, panic_probe as _};
 
 bind_interrupts!(struct Irqs {
-    ADC1 => hal::adc::AdcHandler;
+    ADC1 => InterruptHandler<ADC1>;
 });
-
-#[used]
-#[no_mangle]
-static KEEP_ADC: unsafe extern "C" fn() = ADC1;
 
 #[embassy_executor::main]
 async fn main(_spawner: Spawner) {
-    let p = hal::init(hal::config::Config::default());
+    let mut config = Config::default();
+    config.clock_cfg.sirc.fro_lf_div = Div8::from_divisor(1);
+
+    let p = hal::init(config);
 
     defmt::info!("ADC interrupt Example");
-
-    unsafe {
-        init_adc_pins();
-    }
 
     let adc_config = LpadcConfig {
         enable_in_doze_mode: true,
@@ -41,44 +38,28 @@ async fn main(_spawner: Spawner) {
         trigger_priority_policy: TriggerPriorityPolicy::ConvPreemptImmediatelyNotAutoResumed,
         enable_conv_pause: false,
         conv_pause_delay: 0,
-        fifo_watermark: 0,
         power: PoweredClock::NormalEnabledDeepSleepDisabled,
         source: AdcClockSel::FroLfDiv,
         div: Div4::no_div(),
     };
-    let adc = hal::adc::Adc::<hal::adc::Adc1>::new(p.ADC1, adc_config);
+    let mut adc = Adc::new_async(p.ADC1, p.P1_10, Irqs, adc_config).unwrap();
 
     adc.do_offset_calibration();
     adc.do_auto_calibration();
-
-    let mut conv_command_config = adc.get_default_conv_command_config();
-    conv_command_config.channel_number = Adch::SelectCorrespondingChannel8;
-    conv_command_config.conversion_resolution_mode = Mode::Data16Bits;
-    adc.set_conv_command_config(1, &conv_command_config);
-
-    let mut conv_trigger_config = adc.get_default_conv_trigger_config();
-    conv_trigger_config.target_command_id = Tcmd::ExecuteCmd1;
-    conv_trigger_config.enable_hardware_trigger = false;
-    adc.set_conv_trigger_config(0, &conv_trigger_config);
+    adc.set_resolution(Mode::Data16Bits);
 
     defmt::info!("ADC configuration done...");
-
-    adc.enable_interrupt(0x1);
-
-    unsafe {
-        hal::interrupt::ADC1.enable();
-    }
-
-    unsafe {
-        cortex_m::interrupt::enable();
-    }
+    let mut ticker = Ticker::every(Duration::from_millis(100));
 
     loop {
-        adc.do_software_trigger(1);
-        while !adc.is_interrupt_triggered() {
-            // Wait until the interrupt is triggered
+        ticker.next().await;
+        match adc.read().await {
+            Ok(value) => {
+                defmt::info!("ADC value: {}", value);
+            }
+            Err(e) => {
+                defmt::error!("ADC read error: {:?}", e);
+            }
         }
-        defmt::info!("*** ADC interrupt TRIGGERED! ***");
-        //TBD need to print the value
     }
 }